removing shopping cart
